forecast-service/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.24.2
scipy==1.11.4

# Time Series Models
statsmodels==0.14.0
prophet==1.1.5

# Machine Learning / Deep Learning
tensorflow==2.14.0
scikit-learn==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Job Scheduling
schedule==1.2.1
apscheduler==3.10.4

# Utilities
joblib==1.3.2
pickle5==0.0.11